+++ src/test/java/pages/booking/StaysHomePage.java
@@ -4,7 +4,7 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import steps.SimpleSteps;
+import steps.booking.SimpleSteps;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
