+++ src/test/java/runners/BookingRunner.java
@@ -2,9 +2,9 @@ package runners;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
-import tests.ChangeColorTest;
-import tests.HighestPriceTest;
-import tests.LowestPriceTest;
+import tests.booking.ChangeColorTest;
+import tests.booking.HighestPriceTest;
+import tests.booking.LowestPriceTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({HighestPriceTest.class, LowestPriceTest.class, ChangeColorTest.class})
