+++ src/test/java/steps/booking/ExtendedSteps.java
@@ -1,4 +1,4 @@
-package steps;
+package steps.booking;
 
 public class ExtendedSteps {
 }
