+++ src/test/java/tests/booking/ChangeColorTest.java
@@ -1,4 +1,4 @@
-package tests;
+package tests.booking;
 
 import driver.GetDriver;
 import org.junit.After;
@@ -14,7 +14,7 @@ import pages.booking.StaysSearchResultsPage;
 import settings.DriverConfig;
 import settings.ScreenMode;
 import steps.GeneralSteps;
-import steps.SimpleSteps;
+import steps.booking.SimpleSteps;
 
 import java.util.concurrent.TimeUnit;
